Every StructDescriptor is now a 'MapStructDescriptor'.
