+++ vitrivr-engine-core/src/main/kotlin/org/vitrivr/engine/core/model/descriptor/struct/metadata/TemporalMetadataDescriptor.kt
@@ -3,7 +3,6 @@ package org.vitrivr.engine.core.model.descriptor.struct.metadata
 import org.vitrivr.engine.core.model.descriptor.Attribute
 import org.vitrivr.engine.core.model.descriptor.AttributeName
 import org.vitrivr.engine.core.model.descriptor.DescriptorId
-import org.vitrivr.engine.core.model.descriptor.struct.MapStructDescriptor
 import org.vitrivr.engine.core.model.descriptor.struct.StructDescriptor
 import org.vitrivr.engine.core.model.metamodel.Schema
 import org.vitrivr.engine.core.model.retrievable.RetrievableId
@@ -22,7 +21,7 @@ class TemporalMetadataDescriptor(
     override val retrievableId: RetrievableId?, //retrievable Id must come first, due to reflection
     values: Map<AttributeName, Value<*>?>,
     override val field: Schema.Field<*, TemporalMetadataDescriptor>? = null
-) : MapStructDescriptor<TemporalMetadataDescriptor>(id, retrievableId, SCHEMA, values, field) {
+) : StructDescriptor<TemporalMetadataDescriptor>(id, retrievableId, SCHEMA, values, field) {
 
     companion object {
         /** The field schema associated with a [TemporalMetadataDescriptor]. */
