+++ vitrivr-engine-core/src/main/kotlin/org/vitrivr/engine/core/model/descriptor/struct/metadata/source/FileSourceMetadataDescriptor.kt
@@ -3,7 +3,6 @@ package org.vitrivr.engine.core.model.descriptor.struct.metadata.source
 import org.vitrivr.engine.core.model.descriptor.Attribute
 import org.vitrivr.engine.core.model.descriptor.AttributeName
 import org.vitrivr.engine.core.model.descriptor.DescriptorId
-import org.vitrivr.engine.core.model.descriptor.struct.MapStructDescriptor
 import org.vitrivr.engine.core.model.descriptor.struct.StructDescriptor
 import org.vitrivr.engine.core.model.metamodel.Schema
 import org.vitrivr.engine.core.model.retrievable.RetrievableId
@@ -22,7 +21,7 @@ class FileSourceMetadataDescriptor(
     override val retrievableId: RetrievableId?,
     values: Map<AttributeName, Value<*>?>,
     override val field: Schema.Field<*, FileSourceMetadataDescriptor>? = null
-) : MapStructDescriptor<FileSourceMetadataDescriptor>(id, retrievableId, SCHEMA, values, field) {
+) : StructDescriptor<FileSourceMetadataDescriptor>(id, retrievableId, SCHEMA, values, field) {
 
     constructor(id: DescriptorId, retrievableId: RetrievableId?, path: Value.String, size: Value.Long, field: Schema.Field<*, FileSourceMetadataDescriptor>) :
             this(id, retrievableId, mapOf("path" to path, "size" to size), field)
