+++ vitrivr-engine-module-features/src/main/kotlin/org/vitrivr/engine/module/features/feature/averagecolorraster/RasterDescriptor.kt
@@ -3,7 +3,7 @@ package org.vitrivr.engine.module.features.feature.averagecolorraster
 import org.vitrivr.engine.core.model.descriptor.Attribute
 import org.vitrivr.engine.core.model.descriptor.AttributeName
 import org.vitrivr.engine.core.model.descriptor.DescriptorId
-import org.vitrivr.engine.core.model.descriptor.struct.MapStructDescriptor
+import org.vitrivr.engine.core.model.descriptor.struct.StructDescriptor
 import org.vitrivr.engine.core.model.metamodel.Schema
 import org.vitrivr.engine.core.model.retrievable.RetrievableId
 import org.vitrivr.engine.core.model.types.Type
@@ -20,7 +20,7 @@ class RasterDescriptor(
     override var retrievableId: RetrievableId?,
     values: Map<AttributeName, Value<*>?>,
     override val field: Schema.Field<*, RasterDescriptor>? = null
-) : MapStructDescriptor<RasterDescriptor>(id, retrievableId, LAYOUT, values, field) {
+) : StructDescriptor<RasterDescriptor>(id, retrievableId, LAYOUT, values, field) {
 
     companion object {
         private val LAYOUT = listOf(
